♻️ formatting 함수로 이동
