+++ src/utils/formatting.ts
@@ -52,3 +52,16 @@ export function formatDate(date: number): string {
 export function capitalize(word: string) {
   return word.charAt(0).toUpperCase() + word.slice(1);
 }
+
+export function formatDateAndTime(timestamp: string | number): string[] {
+  const date = new Date(timestamp);
+
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+
+  const hour = String(date.getHours()).padStart(2, "0");
+  const minute = String(date.getMinutes()).padStart(2, "0");
+
+  return [`${year}.${month}.${day}`, `${hour}:${minute}`];
+}
